Mapper-evolution functions cleaned (and to records).

+++ src/main/java/it/units/malelab/jgea/problem/classification/DataUtils.java
@@ -34,7 +34,7 @@ public class DataUtils {
           .toList();
       subset.addAll(
           currentSubset.stream()
-              .skip(currentSubset.size() / n * i)
+              .skip((long) currentSubset.size() / n * i)
               .limit(currentSubset.size() / n).toList());
     });
     return subset;
